+++ pioreactor/automations/dosing/pid_morbidostat.py
@@ -7,7 +7,7 @@ from pioreactor.automations import events
 from pioreactor.automations.dosing.base import DosingAutomationJob
 from pioreactor.config import config
 from pioreactor.exc import CalibrationError
-from pioreactor.utils import local_persistant_storage
+from pioreactor.utils import local_persistent_storage
 from pioreactor.utils.streaming_calculations import PID
 
 
@@ -29,7 +29,7 @@ class PIDMorbidostat(DosingAutomationJob):
         assert target_normalized_od is not None, "`target_normalized_od` must be set"
         assert target_growth_rate is not None, "`target_growth_rate` must be set"
 
-        with local_persistant_storage("current_pump_calibration") as cache:
+        with local_persistent_storage("current_pump_calibration") as cache:
             if "media" not in cache:
                 raise CalibrationError("Media pump calibration must be performed first.")
             elif "waste" not in cache: