+++ pioreactor/background_jobs/leader/mqtt_to_db_streaming.py
@@ -83,7 +83,11 @@ class MqttToDBStreamer(LongRunningBackgroundJob):
             for topic_to_table in topics_to_tables
         ]
 
-        self.timer = RepeatedTimer(60, self.write_stats).start()
+        self.timer = RepeatedTimer(
+            60,
+            self.write_stats,
+            job_name=self.job_name,
+        ).start()
 
         self.initialize_callbacks(topics_and_callbacks)