+++ pioreactor/cli/pio.py
@@ -21,7 +21,7 @@ from pioreactor.pubsub import get_from
 from pioreactor.pubsub import post_into_leader
 from pioreactor.utils import JobManager
 from pioreactor.utils import local_intermittent_storage
-from pioreactor.utils import local_persistant_storage
+from pioreactor.utils import local_persistent_storage
 from pioreactor.utils.networking import is_using_local_access_point
 from pioreactor.utils.timing import current_utc_timestamp
 
@@ -210,7 +210,7 @@ def view_cache(cache: str) -> None:
         cacher = local_intermittent_storage
 
     elif Path(f"{persistant_dir}/{cache}").is_dir():
-        cacher = local_persistant_storage
+        cacher = local_persistent_storage
 
     else:
         click.echo(f"cache {cache} not found.")
@@ -240,7 +240,7 @@ def clear_cache(cache: str, key: str, as_int: bool) -> None:
         cacher = local_intermittent_storage
 
     elif Path(f"{persistant_dir}/{cache}").is_dir():
-        cacher = local_persistant_storage
+        cacher = local_persistent_storage
 
     else:
         click.echo(f"cache {cache} not found.")