+++ pioreactor/tests/test_pumps.py
@@ -22,7 +22,7 @@ from pioreactor.pubsub import publish
 from pioreactor.pubsub import subscribe
 from pioreactor.pubsub import subscribe_and_callback
 from pioreactor.utils import local_intermittent_storage
-from pioreactor.utils import local_persistant_storage
+from pioreactor.utils import local_persistent_storage
 from pioreactor.utils import timing
 from pioreactor.whoami import get_unit_name
 
@@ -34,13 +34,13 @@ def pause(n=1):
 
 
 def setup_function():
-    with local_persistant_storage("current_pump_calibration") as cache:
+    with local_persistent_storage("current_pump_calibration") as cache:
         cache["media"] = encode(
             structs.MediaPumpCalibration(
                 calibration_name="setup_function",
                 curve_data_=[1.0, 0.0],
                 curve_type="poly",
-                recorded_data={"x": [], 'y': []},
+                recorded_data={"x": [], "y": []},
                 dc=60,
                 hz=100,
                 created_at=datetime(2010, 1, 1, tzinfo=timezone.utc),
@@ -53,7 +53,7 @@ def setup_function():
                 calibration_name="setup_function",
                 curve_data_=[1.0, 0.0],
                 curve_type="poly",
-                recorded_data={"x": [], 'y': []},
+                recorded_data={"x": [], "y": []},
                 dc=60,
                 hz=100,
                 created_at=datetime(2010, 1, 1, tzinfo=timezone.utc),
@@ -66,7 +66,7 @@ def setup_function():
                 calibration_name="setup_function",
                 curve_data_=[1.0, 0.0],
                 curve_type="poly",
-                recorded_data={"x": [], 'y': []},
+                recorded_data={"x": [], "y": []},
                 dc=60,
                 hz=100,
                 created_at=datetime(2010, 1, 1, tzinfo=timezone.utc),
@@ -92,7 +92,7 @@ def test_pump_io() -> None:
 def test_pump_fails_if_calibration_not_present() -> None:
     exp = "test_pump_fails_if_calibration_not_present"
 
-    with local_persistant_storage("current_pump_calibration") as cache:
+    with local_persistent_storage("current_pump_calibration") as cache:
         del cache["media"]
         del cache["alt_media"]
         del cache["waste"]
@@ -211,7 +211,7 @@ def test_pump_can_be_interrupted() -> None:
         calibration_name="setup_function",
         curve_data_=[1.0, 0.0],
         curve_type="poly",
-        recorded_data={"x": [], 'y': []},
+        recorded_data={"x": [], "y": []},
         dc=100,
         hz=100,
         created_at=datetime(2010, 1, 1, tzinfo=timezone.utc),
@@ -258,7 +258,7 @@ def test_pumps_can_run_in_background() -> None:
         calibration_name="setup_function",
         curve_data_=[1.0, 0.0],
         curve_type="poly",
-        recorded_data={"x": [], 'y': []},
+        recorded_data={"x": [], "y": []},
         dc=60,
         hz=100,
         created_at=datetime(2010, 1, 1, tzinfo=timezone.utc),
@@ -317,13 +317,13 @@ def test_waste_pump_cant_run_when_media_circulation_is_running() -> None:
 def test_media_circulation_will_control_media_pump_if_it_has_a_higher_flow_rate() -> None:
     exp = "test_media_circulation_will_control_media_pump_if_it_has_a_higher_rate"
 
-    with local_persistant_storage("current_pump_calibration") as cache:
+    with local_persistent_storage("current_pump_calibration") as cache:
         cache["media"] = encode(
             structs.MediaPumpCalibration(
                 calibration_name="setup_function",
                 curve_data_=[10.0, 0.0],
                 curve_type="poly",
-                recorded_data={"x": [], 'y': []},
+                recorded_data={"x": [], "y": []},
                 dc=60,
                 hz=100,
                 created_at=datetime(2010, 1, 1, tzinfo=timezone.utc),
@@ -336,7 +336,7 @@ def test_media_circulation_will_control_media_pump_if_it_has_a_higher_flow_rate(
                 calibration_name="setup_function",
                 curve_data_=[1.0, 0.0],
                 curve_type="poly",
-                recorded_data={"x": [], 'y': []},
+                recorded_data={"x": [], "y": []},
                 dc=60,
                 hz=100,
                 created_at=datetime(2010, 1, 1, tzinfo=timezone.utc),
@@ -352,13 +352,13 @@ def test_media_circulation_will_control_media_pump_if_it_has_a_higher_flow_rate(
 def test_media_circulation_will_control_media_pump_if_it_has_a_lower_flow_rate() -> None:
     exp = "test_media_circulation_will_control_media_pump_if_it_has_a_lower_flow_rate"
 
-    with local_persistant_storage("current_pump_calibration") as cache:
+    with local_persistent_storage("current_pump_calibration") as cache:
         cache["media"] = encode(
             structs.MediaPumpCalibration(
                 calibration_name="setup_function",
                 curve_data_=[0.15, 0.0],
                 curve_type="poly",
-                recorded_data={"x": [], 'y': []},
+                recorded_data={"x": [], "y": []},
                 dc=60,
                 hz=100,
                 created_at=datetime(2010, 1, 1, tzinfo=timezone.utc),
@@ -371,7 +371,7 @@ def test_media_circulation_will_control_media_pump_if_it_has_a_lower_flow_rate()
                 calibration_name="setup_function",
                 curve_data_=[1.0, 0.0],
                 curve_type="poly",
-                recorded_data={"x": [], 'y': []},
+                recorded_data={"x": [], "y": []},
                 dc=60,
                 hz=100,
                 created_at=datetime(2010, 1, 1, tzinfo=timezone.utc),
@@ -386,7 +386,7 @@ def test_media_circulation_will_control_media_pump_if_it_has_a_lower_flow_rate()
 
 def test_media_circulation_works_without_calibration_since_we_are_entering_duration() -> None:
     exp = "test_media_circulation_works_without_calibration_since_we_are_entering_duration"
-    with local_persistant_storage("current_pump_calibration") as cache:
+    with local_persistent_storage("current_pump_calibration") as cache:
         del cache["media"]
         del cache["waste"]